improve Makefile for FreeBSD

+++ Makefile
@@ -45,7 +45,7 @@ install: st
 	sed "s/VERSION/$(VERSION)/g" < st.1 > $(DESTDIR)$(MANPREFIX)/man1/st.1
 	chmod 644 $(DESTDIR)$(MANPREFIX)/man1/st.1
 	tic -sx st.info
-	mv st.info.cdb ~/.terminfo.cdb
+	if [ -e st.info.cdb ]; then mv st.info.cdb ~/.terminfo.cdb; fi
 	@echo Please see the README file regarding the terminfo entry of st.
 
 uninstall:
